feat(scans)!: added return_to_start kwarg to all scans. If set to true, the scan returns to start.
BREAKING CHANGE: In the past, this was always set to true. Now, the default will depend on the relative flag. Relative scans return to start to avoid shifts for repeated scans. Absolute scans will not return to start. However, this behaviour can be overridden by providing the additional return_to_start kwarg or subclassing the scan base.

closes #332